backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
pydantic>=2.6.4
pydantic-settings==2.2.1
email-validator>=2.2.0
python-dotenv>=1.0.1